Add a `copyable` concern
This allows us to make a copy of an edition with the relations and
immutable attributes intact, as well as setting the state to `draft`
(the initial state of all editions). In the next commit, we’ll replace
the call to `create_new_content_block_edition_for_document` in the
`Publishable` concern to use this new method.

I’ve added some comments in a Gherkin style, as there is a lot of
behaviour being tested and the comments make it a bit clearer as to
what is going on.