Document government override feature
This feature was added to Whitehall to enable GDS editors to specify a particular government that should appear on the historical content banner for a document.

This feature is likely to be used mainly for content that was published shortly after a change of government.

Also add a footnote to note that all content types are linked to the government in power when they were first published.