Refactor flaky test
In some limited scenarios the sentry tags test was erroring with an
error “please make sure the SDK is initialized for testing” coming from
Sentry. This seemed to be some issue with test ordering that I couldn’t
track down.

This PR changes the test to stub the call to `set_tags` and check this
is called in a test controller that subclasses the base controller. This
is less inclined to be flaky as we’re testing only the internal
behaviour of our code, rather than relying on Sentry to be set up
correctly and then testing the behaviour of the Sentry gem itself (which
we should trust behaves correctly anyway).