Default to `EMERGENCY_BANNER_REDIS_URL` in controller
This is a new environment variable we've currently added to Integration
to allow us to switch Whitehall to a new Redis instance.

For environments that don't have this variable set yet (i.e. Staging and
Production) we fallback on the default `REDIS_URL`.

This is temporary - once we've moved all environments over to the new
URL we can remove this default.

`mock_env` is shamelessly copied from
[here](https://gist.github.com/jazzytomato/79bb6ff516d93486df4e14169f4426af)
and allows us to specify some environment variables that won't persist
between tests.

I'd probably extract this to a helper for future use but this is a
temporary change that we'll remove fairly soon.