Add task to publish empty landing pages
This is basically a copy of the PublishFinders code, including its
tests.

The idea is that we'll publish some of these landing page content items
just so that their routes are declared and point to the correct frontend
(which is going to be frontend). The details and links of the content
item will initially be added by the frontend based on hardcoded YAML
files. Later, we'll work out a CMS for these content items, which may be
Whitehall, or may be something else entirely (e.g. a third party CMS
like Sanity of Contentful [other CMSs are available])