Read TAXONOMY_CACHE_REDIS_URL if set
This is a new environment variable we've currently only added to
Integration as of [this
commit](https://github.com/alphagov/govuk-helm-charts/pull/2598/commits/501163523cee534667548ab9b288060d2f595ae4),
and points to a separate database for Whitehall's own Redis DB.

For environments that don't have this variable set yet (i.e. Staging and
Production) we fallback on the default `REDIS_URL`.

This is temporary - once we've moved all environments over to the new
URL we can remove this default.