Move type match validation logic to publication
Since the National Statistics have been renamed to Accredited Official
Statistics, we've had a few Zendesk tickets of users not realizing that
there was a mismatch between the publication and announcement type.
Previously the validation of the type match was done on announcement, but
never surfaced on the publication model at the point of saving. This gave
the users a generic "Something went wrong" error.

The validation logic is primarily relevant when the publication gets saved,
as its type must match that of its announcement. Users tend to create the
publication off of the announcement. With the validation in the publication model,
we can be more expressive and set the error on the publication_type field,
which gives us a nice error summary link to the field.

We still want the validation to run on announcement, which we can achieve by
using a validates_associated, and surface the publication errors, with some
customization.