Add guidance text to connect document page
The validation added in previous commits works fine for saving and editing. Nonetheless,
users also have the option to associate a pre-existing publication to an announcement, with the
"Connect document" feature. This was also reported in Zendesk as a source of confusion, since the
filter is opinionated about the publication and announcements types matching, but nothing on the page
communicates that to the user.
We've added a text inset to indicate that this is the case.