Move `stub_request_for_schema` to a helper
We were getting flaky tests when using `stub_request_for_schema` in two
places because the helper function was defined in two files, but the
implementation was slightly different. This unifies the behaviour,
returning the stub schema in both places, whilst also moving the helper
code to a module, which we include when we need it.