Fix error messaging when changing statistics announcement type with a published publication
Autosaving as well as validating publication now surfaces the error message for published publication twice.
To handle this, I've tried putting on: :create back onto validates_associated for publication but this started failing to catch the scenario where we are trying to connect a document of a wrong type. In the end, the only way to satisfy all test cases, was to force a validate: false on the autosave, and rely entirely on validates_associated