Add missing tests for SelectWithSearchHelper
In the next commit, we're bumping the Choices.js library to a new
version, which has different behaviour to before. There are
currently no tests defining the expected input/output of the
GovukPublishingComponents::Presenters::SelectWithSearchHelper,
so we must add these first - we can then see what code needs
changing to accommodate the changes in Choices.js.