Do not validate a `Feature` when ending it
We've had a report of a publisher being unable to unpublish a
document because it has 'features' which do not validate. The
feature in question is missing an image, which was added as a
validation rule [11 months ago](https://github.com/alphagov/whitehall/blob/1c29bf277c69a0a134d44aa6a9f5c5a29cde63dc/app/models/feature.rb#L65-L66).
The featuring predates the validation.

When we're calling `.end!` on a feature, we really shouldn't care
whether or not the feature is valid, since it's literally being
ended. There is precedent for that, e.g. in the code that triggers
the `.end!` method, which already calls `.save!(validate: false)`
on the edition itself:
https://github.com/alphagov/whitehall/blob/1dedd82dffa8680d2db4b741675fa68d2243fc3f/app/services/edition_unpublisher.rb#L25-L26

Trello: https://trello.com/c/LFCD7dMp/3011-superseded-by-another-page-unpublish-content-request