Replace the body textarea with a Monaco editor
This should provide a much better editing experience for YAML files. It
gives us syntax highlighting, code folding, line numbers, and a whole
bunch of other features.

It does not yet provide schema validation, or schema-powered
autocomplete, but that's a possible improvement for later.

We're using ecmascript modules to load the third party javascript here,
rather than installing it with npm. This is just because bundling the
code and making the rails asset pipeline aware of it is a yak shave.
Hopefully the YAML editor won't be a long term component, so this will
be a temporary oddity.