Have landing page extend document collection
The theory here is that often, landing pages effectively own collections of documents, and it's nice for those documents to be linked together in some way.

By extending document collection, we get some UI that allows us to add groups of documents, and publish them as links to the landing page.

Documents links ought to be reversed by publishing API, so the documents themselves should link back to the landing page.