Return Updated info on Host Editions
This attempts to find a user for the `last_edited_by_editor_id` for a
piece of content, if it exists.

Currently, we’re querying the database for this information, which will
only work if a user has Whitehall access. This isn’t ideal, and we’ve
added a card to the backlog to look into how else we can access this
information (possibly via Signon). However, this is a bigger piece of
work that we’ll look at tackling further down the line.