Make republishable_content_types less clever
Using `Edition.descendants` makes some assumptions about the inheritance
tree of Edition which might not always be true.

Specifically, at the moment the only classes that get published are
leaf nodes in the tree - classes without any descendants. In another
branch, I would like to make a new LandingPage class which inherits
DocumentCollection. Both LandingPage and DocumentCollection should be
publishable, so they should both appear in the republishable documents
list.

I don't think there's a neat way to account for this, and we have to
hardcode the lists of document types in the tests anyway, so I think
it's probably fine to just... hardcode them in the production code.

This also removes the need for us to keep calling eager_load in the
tests, which was a bit smelly because it indicates that the production
code would give incorrect results if the classes in question hadn't been
loaded for some reason.

I've also made non_editionable_content_types hardcode its list, even
though the heuristic it's using seems to be more accurate / less likely
to break with changes to the inheritance structure.

I've removed the non_editionable_content_types test, because with both
sides hardcoded it's a bit meaningless.