Make republishable_content_types less clever
Using `Edition.descendants` makes some assumptions about the inheritance
tree of Edition which might not always be true.

Specifically, at the moment the only classes that get published are
leaf nodes in the tree - classes without any descendants. In another
branch, I would like to make a new LandingPage class which inherits
DocumentCollection. Both LandingPage and DocumentCollection should be
publishable, so they should both appear in the republishable documents
list.

The nice thing about the old approach was that if we add models in the
future, they'd be automatically included (but the tests would fail
because they had a hardcoded list which would need updating).

I've switched it around, so that the production code hardcodes the
models (which is nice and simple and easy to reason about), but the
tests look at the descendants. We'll still get a failing test if we add
a model that's missing from the hardcoded list, because the tests will
spot the new model.