Rake task to identify all attachment data with all attachments only on superseded
This did not run to completion.
All non-deleted attachments on all superseded editions are valid if they have replacements in asset manager.
This rake task gives us an idea on how many have replacements set in wh, that we should further interrogate against AM. We can either fix these or delete them.

For the ones that do NOT have replacements in WH either, that is an "impossible" state and we can delete them.