set filter params logic for content blocks
This mimics the default behaviour when viewing all
Documents in Whitehall:
* if there are no filters provided, then the
  user's organisation is set as the lead
  organisation
* otherwise we take from the session so that their
  search is not lost between page navigations.

This does make other integration and feature tests
a bit complicated - as the first time a user goes
to the page their organisation will be set as a
filter. In the feature tests I'm checking that and
then setting to 'All organisations' to ensure the
scope of other tests is set to specific filters.