WIP: delete obsolete translations when reverting back to English
I wasn't sure we'd need this extra logic, given that the
reverse logic doesn't need it (switching from English to a foreign
language updates the primary locale and changes the only translation
locale from English to the chosen language, which is exactly the
behaviour we want in reverse).

However, switching from foreign back to English keeps the foreign
translation. So we do need to manually clean up the old
translations.