Add safeguard against accidentally deleting multiple translations
As outlined in the previous commit, there should never be a
situation where the primary locale has changed AND there are
multiple translations. It should simply go from (for example)

- `cy` translation (primary)
- `cy` and `en` translations (`en` being the new primary)
- `en` translation (`cy` deleted as part of the before_validation
  callback)

That said, Whitehall can be full of surprises, and it's probably
better that we preserve any translations if we DO find ourselves
in a state where there are multiple translations and a changing
primary locale. It is better that the department raise a support
ticket with us so we can look into it further and step in
manually, than to make the wrong assumption and delete their
translations.