Use image kind config to validate image data
... instead of hardcoding sizes.

Initially this just replaces the various hardcoded `size: [960, 640]` parameters with values from image_kind_config on the model, which is hardcoded to be "default".

When we have multiple different image kinds (e.g. hero images, featured images, default images), we can make the image_kind method into an active record attribute and store the kind of image in the database instead of hardcoding it. This will allow different ImageData objects to have different valid sizes, depending on their kind.

This commit should not change any behaviour, so many tests (e.g. the ones in image_data_test.rb) are unchanged.