Refactor checking for required variants
We can get a list of the active versions from the mounted ImageUploader (which is mounted as file). The method we want is protected, but we can usefully wrap it up.

This reduces a bit of duplication compared to the bitmap? ternary, which was doing the same thing as the `if: bitmap?` conditions we have on all the variants.