Allow assets to have any configured variant name
Previously we were duplicating the hardcoded list of image variants (s960, s712 etc.) in ImageUploader and Asset. Presumably only variants appearing in both lists would be uploaded as assets.

This pulls the list of versions / variants from the image_kinds.yml config file. Once ImageUploader is similarly refactored, this will remove a bit of duplication. It will also allow versions from other (non-default) kinds of images (e.g. hero images) in future.

It's now important that version names in the config are globally unique, so I've added a test that checks they are.