Use image versions from config for image uploader
... instead of hardcoding them.

Carrierwave makes this a little bit tricky - because version is a class method, we can't call it conditionally based on instance methods (like model.image_kind), because there is no instance when the method is called.

Carrierwave provides an "if: " key to work around this, which previously we were just using to check if the uploaded file is a bitmap.

By passing a lambda in instead, we can access both the newly uploaded file _and_ the model the uploader is mounted on. In a previous commit we added the image_kind method to all such models.

This means we can add every possible version (for all different kinds of image) to ImageUploader, but have them check in their if condition that the model they're attached to has a relevant kind.

At this point this is all a bit abstract, because there's only one image kind in the config ("default"). But when we later want other kinds of image (e.g. hero images), this approach should be flexible enough to support that.