Add JS to copy embed code
This finds any row in a summary card/list component with a
`data-module=“copy-embed-code”` and adds a `dd` element with a link to
copy the code. If clicked, this gets the value of summary list row and
copies it to the clipboard.

The modern way of copying to the clipboard (via
`navigator.clipboard.writeText`) isn’t supported in all browsers and
causes issues with testing, so we’re doing it the “old-fashioned way” by
creatinf an invisible textarea, populating it with the text, then
selecting and coping it it via `execCommand`.

I’ve also added some tweaks to the Jasmine setup, so we can keep our
Jasmine specs within the engine.