Link to draft host document for content block preview
We would like to link to the draft store so that
users can see the changes they've just made to the
block in host documents. As this table component
is also used on the 'view' page I am passing in a
boolean to determine whether to show draft or live
frontend.