open host document links in new tab
Although opening a new tab is not ideal for
accessibility, in this case we think it is the
preferred behaviour - opening a new tab can be
recommended if a user is in the middle of a for or
would lose context of what they are working on if
not taken to a new tab. In the case of these host
documents, if the user is previewing changes they
will be in the middle of the form, if they are
just viewing the content block it could be
confusing to be taken to gov.uk and out of Content
Block Manager.

There is precedent across Whitehall GDS for
opening in new tabs when needed, and the
mitigation is to include a message for screen
readers declaring that the link opens in a new
tab. We can test how this approach works for our
users and change if needed.