paginate Content Block search results
This uses a similar approach as the Whitehall
app's Edition Filter but a bit simplified - we
don't currently need to set the page size
externally so I've set it to be the shared default
of 15 for now.

I've also removed mocking in an integration test
on the controller, because (when possible) we want
integration tests to cover the end-to-end
logic and behaviour rather than stubbing or mocking.