Allow bulk republishing all non-editionables
When working on GraphQL in Publishing API, we found that content like
role appointments were not pushed to Publishing API when running the
Whitehall seeds, nor when subsequently running the UI task to republish
all documents. We needed role appointments to be pushed to Publishing
API in order to test out queries for the data that supports the prime
minister page, and later the ministers index page (though the seed data
are limited, so they would only be so useful here)

We found that we could do this by going into each role appointment and
re-saving it, however this is quite a manual task, and one that we had
to work out how to do. The reason republishing all documents didn't push
this content is that role appointments are non-editionable content, and
documents in Whitehall are parents of editions. Under the hood,
republishing all documents republishes all documents' editions to
Whitehall. Non-editionable content is by definition excluded

This therefore adds support for quickly republishing all non-editionable
content (excluding individual presenter-based pages, which will be
handled in a subsequent commit) to Publishing API, aiding such local
development needs as motivated this work

As outlined in an earlier commit, we could separately add support for
republishing all publishable content in a single task, but for now I'm
opting for a balance of control and efficiency. These tasks would be
expensive in live/non-development environments where the amount of
publishable content is very large, so it's preferable to be able to
limit the scope of republishing tasks (while also providing quick ways
of pushing all content in a small number of steps)