Allow bulk republishing all individual pages
These are pages like the ministers index and world index. They are
represented by presenter classes in Whitehall. Note that while the UI
says "All individual pages", it's only the seven supported pages as
outlined in the republishing index (the top section). We might want to
clarify this, but this is a minimal implementation for now