Clarify scope of bulk republishing all documents
We have a bulk republishing method that republishes all documents. We
might have hoped that running this task would push all publishable
content to Publishing API, however the concept of documents in Whitehall
is not quite equivalent to documents in Publishing API. Like in
Publishing API, documents in Whitehall are parents of editions, however
Whitehall has a concept of non-editionable content, i.e. content that
only has a single version with no support for drafts. Republishing all
documents therefore only pushes editionable content to Publishing API,
since that's the extent of what documents represent in Whitehall. We
could change this so that all publishable content is pushed by this
method, but that might confuse colleagues who are used to the Whitehall
definition of documents

This therefore clarifies the scope of bulk republishing all documents,
aiding those of us who are used to the Publishing API definition of
documents, or who work across both apps and may find the term ambiguous

Upcoming commits will add equivalent methods for bulk republishing the
seven individual pages listed in the republishing UI, as well as all
other non-editionable content types. We could potentially add a separate
method to republish all publishable content later