Add permitted_uses to image kind config
This means we don't have to hardcode the fact that only default images are allowed to be embedded in govspeak. If we later have other kinds of images which are allowed to be embedded in govspeak, that could be represented in config too.

The permitted_uses array could be used in future for other situations where only some kinds of images are allowed to be used in a particular circumstance - for example in featured images, lead images, person images, etc.