Table component supports sorting and pagination
Each sort / page link also has an anchor link to the table (the ID is
the `TABLE_ID` constant. This ensures the user doesn’t lose focus when
sorting/paging. I did consider some fancy JS solution, but this seems to
work pretty well for now. We can always iterate later if we need to.