Update styling
When applying `sortable: true` when using the Table component, the
design changes quite a lot from the default look and feel of the
stock GOV.UK table. (See https://components.publishing.service.gov.uk/component-guide/table#with_sortable_head)
To get around this, we’re adding some of the styling in the component
just to the headers so the look and feel is more akin to the designs
we have.