Remove 'organisation_attachments_report' rake task
This was added in #4970 to prepare organisations for meeting new
accessibility regulations. It output a CSV report of their
"non-accessible attachments" so that they could take steps to
improve them.

As per https://trello.com/c/BSfJQeUu/1162-report-listing-nonaccessible-attachments-on-govuk:

> In the first instance, this is likely to be a rake task, run by developers on demand. It should take organisations as a parameter, or run for all orgs.
> That'll let us test the user needs better before further work.

It doesn't look as though any further work was forthcoming.

The generated file on the Trello card was ~2.5mb in size, so if
needed we could bring this rake task back, and just output to
STDOUT.