Remove old 'export:document_list' rake task
This was added 12 years ago in commit 24636257dd22cfd86e04417e8519a74054e7e589.
It's unclear when it was last used, but there's no obvious use
case for it now.

Documents can be exported as a CSV through Whitehall's UI on the
document search page. Anything more detailed can be done by a
dev on the rails console.