Update Govspeak regex Rake task
This Rake task finds content with Govspeak matching a given regex. When
trying to use this to get instances of trailing space within address
blocks, I encountered a few issues. Relevant code is updated to deal
with additional possible data, and a comment is added with some guidance
on how to deal with likely issues. This should prevent other devs having
to work out how to get around these issues in future

The link to the dev docs is subject to a change in the URL, which is
being proposed here: https://github.com/alphagov/govuk-developer-docs/pull/4900