Republish documents after updating access limits
I was just burned by this not happening. I updated a bunch of editions to add number 10 to the access limited organisations, only to discover that number 10 still couldn't access them.

On inspection, I worked out that the access limits had been updated in Whitehall, but not content store.

Republishing the documents should ensure that they make it to the draft / live content stores any time their access limits are updated.