Fix tests for carrierwave upgrade
Carrierwave has introduced a new temporary cache ID to its carrierwave_file names to force trigger a file update if users upload a file with the same name. This has then broken our asset comparison checks since we rely on these filenames to check whether or not the assets have been updated in assets manager, and the filename no longer matches. We are switching to use filename instead, which is the same value we send to Asset Manager for a more consistent check

We could use these temporary cache IDs to our advantage to differentiate between files with the same name, but unfortunately this gets stripped upon saving, and hence will not be available after commit. Not a consistent value to depend on. It may also affect readability of the generated URL filepaths

Simplifying logic of asset checking

Instead of relying on splitting, we could just use the filename that we actually pass through to asset manager to do the checks