Validate blocks with ActiveModel validators
Initially this is just the basic scaffolding - a factory for building
"blocks" from ruby hashes, and a base block class which just validates
that the block has a type.

Each block has a present_for_publishing_api method, which turns it back
into a hash in the format publishing-api wants.

The LandingPage::Body class validates that none of its child blocks are
`invalid?`, so errors are passed up the chain.

We'll need to implement stricter block specific validation in future
commits, but for now so long as a block is a hash, and has a type it
will be treated as valid.