Do not apply button tracking to elements with button role
We are preventing button tracking for elements that are not buttons but have the button role because there is a risk of duplicate events.

Most non-button elements with the button role are links, and will therefore be tracked via the link tracker module. The link tracker module will send a link event type for links with the button role. We do not need to send additional events via button tracking for such elements.