add error handling for `Net::HTTP.get` to frontend
the `Net::HTTP.get` doesn't seem to throw an error
when the frontend end is down/unavailable.
According to [docs it looks like it should
throw](https://ruby-doc.org/stdlib-2.7.0/libdoc/net/http/rdoc/Net/HTTP.html#class-Net::HTTP-label-Response+Data
) HTTPServiceUnavailable but it doesn't seem to in
practice, instead it returns the raw HTML of the
"Service Temporarily Unavailable" page.

In the case where the call does error, it looks
like we either have a [long list of all possible
errors, or have a catchall `StandardError` rescue
](https://stackoverflow.com/a/11802674) - this
seemed the most practical to add.

We can revisit this in future if it becomes more
of a priority, perhaps looking at the raw HTML
that's returned to determine if it's what we
expect.