Stop validating uniqueness on attachment filenames
We've now had two support tickets relating to existing, published
Whitehall documents that cannot be edited because the 'create
draft edition' workflow fails to pass validation:
- https://govuk.zendesk.com/agent/tickets/5898727
- https://govuk.zendesk.com/agent/tickets/5978126

In these cases, documents had attachments that shared an identical
filename, causing failures like "Validation failed: This
publication already has a file called \"foo.pdf\"" when trying to
create a draft edition.

It's surprisingly difficult to find a workaround for this; we
can't just pass `validate: false` in the Rails console because
of the convoluted edition superseding logic in Whitehall. We have
no 'hook' to pass the validation override into.
The alternative is to delete one of the duplicate attachments,
which is a developer task that is also quite a sensitive operation.

I've tracked the validation logic addition down to this PR from 2013:
https://github.com/alphagov/whitehall/pull/1170

The linked pivotal tracker story explains the purpose:
> So that we can generate nice urls for attachments, e.g.
> (/publications/slug-for-pub/attachments/filename.pdf), we need
> to ensure all attachments on a given attachable are unique.

This reasoning no longer holds true, now that we're serving assets
on the assets domain and not via Whitehall Frontend.

Note that there is a very similar validation rule enforcing
uniqueness on IMAGE files, added in #7491. We're not touching that
rule, which _does_ appear to still be needed, as images can be
embedded in GovSpeak and thus it's important to be able to
confidently refer to one single asset, no duplicates.

Trello: https://trello.com/c/FJuN46qU/3217-technical-fault-with-whitehall