Call `schedule_republishing_workers` on CallForEvidence
In the event that all Sidekiq jobs are dropped, the
`requeue_all_jobs` rake task is intended to recreate all the
schedule-like jobs so that everything continues to be published
on time as expected.

Most editions can be scheduled for publishing as standard, but
editions that pull in the `Edition::HasOpeningAndClosingDates`
module (i.e. Consultation and CallForEvidence) also need to be
scheduled to publish on the opening or closing date. The rake
task was already doing this for Consultations, so should be doing
it for Calls for Evidence too.